Restructuring common scripts (formerly auxiliary.py and twoBytwo_defs.py) and reorganizing truth kinematics scripts.
