+++ truth_kinematics/outputs/dictionaries/cc_threshold_bkg_dict.json
