+++ truth_kinematics/outputs/dictionaries/dirt_bkg_dict.json
