+++ truth_kinematics/outputs/dictionaries/signal_dict.json
