+++ truth_kinematics/outputs/text/signal_event_counts.txt
@@ -1,0 +1,3 @@
+Signal Events (scaled to 2.5e19 POT): 14070.0
+Number of files used to get count: 1023
+Scale factor:2.5
