+++ truth_kinematics/plotting_scripts/plot_pid_background.py
