+++ truth_kinematics/plotting_scripts/plot_signal_hadrons.py
