+++ truth_kinematics/plotting_scripts/plot_stacked_kinematics.py
