+++ truth_kinematics/plotting_scripts/plot_threshold_background.py
