+++ truth_kinematics/plotting_scripts/signal_plots_from_json_dicts.py
