+++ truth_kinematics/signal_characterization.py
