+++ truth_kinematics/signal_kinematics.py
