+++ truth_kinematics/template_kinematics_characterization.py
